+++ src/addons/action-button.js
@@ -22,9 +22,10 @@ inherit(ActionButton, PureComponent, {
     let action = this.send(this.props.action, this.props.value)
 
     if (action && action instanceof Action) {
-      action.onDone(this.props.onDone)
+      action.onOpen(this.props.onOpen)
             .onUpdate(this.props.onUpdate)
             .onCancel(this.props.onCancel)
+            .onDone(this.props.onDone)
             .onError(this.props.onError)
     }
 
@@ -41,6 +42,7 @@ inherit(ActionButton, PureComponent, {
     delete props.tag
     delete props.action
     delete props.value
+    delete props.onOpen
     delete props.onDone
     delete props.onUpdate
     delete props.onCancel